backend/requirements.txt
Flask==3.0.2
flask-cors==4.0.0
mysql-connector-python==8.3.0
Werkzeug==3.0.1
python-dotenv==1.0.1 
gunicorn ==20.1.0
bcrypt==4.0.1   
PyJWT==2.8.0
